+++ herakoi/core.py
@@ -5,6 +5,7 @@ import re
 
 import cv2
 import mediapipe as mp
+import pygame
 
 import tkinter
 import tkinter.filedialog as filedialog
@@ -27,7 +28,10 @@ scrh = root.winfo_screenheight()
 fill = 1.00
 root.withdraw()
 
+pygame.init()
+
 global pressed; pressed = None
+global  screen; screen  = None
 
 # Convert BGR image into HSV
 # -------------------------------------
@@ -80,6 +84,7 @@ def on_press(key):
 class start:
   def __init__(self,image=None,mode='single',port={},video=0,box=2,switch=True,**kwargs):
     global pressed
+    global screen
 
     self.listener = keyboard.Listener(on_press=on_press)
     self.listener.start()  
@@ -121,15 +126,21 @@ class start:
     self.imgfull = kwargs.get('fullscreen',False)
     self.padfull = kwargs.get('pad',False) if self.imgfull else False
 
+    if self.imgfull:
+      screen = pygame.display.set_mode((scrw,scrh),pygame.FULLSCREEN)
+
     while True:
       self.opvideo = cv2.VideoCapture(video)
       self.opmusic = gethsv(imgpath[imginit])
 
-      if self.imgfull:
-        cv2.namedWindow('mixframe',cv2.WND_PROP_FULLSCREEN)
-        cv2.setWindowProperty('mixframe',cv2.WND_PROP_FULLSCREEN,cv2.WINDOW_FULLSCREEN)
-      else:
+      if not self.imgfull:
         cv2.namedWindow('mixframe',cv2.WINDOW_NORMAL)
+
+    # if self.imgfull:
+    #   cv2.namedWindow('mixframe',cv2.WND_PROP_FULLSCREEN)
+    #   cv2.setWindowProperty('mixframe',cv2.WND_PROP_FULLSCREEN,cv2.WINDOW_FULLSCREEN)
+    # else:
+    #   cv2.namedWindow('mixframe',cv2.WINDOW_NORMAL)
 
       self.mphands = mp.solutions.hands
       self.mpdraws = mp.solutions.drawing_utils
@@ -359,8 +370,13 @@ class start:
           mixframe[                  int(0.01*mixframe.shape[0]):opframe.shape[0]+int(0.01*mixframe.shape[0]),
                    -opframe.shape[1]-int(0.01*mixframe.shape[0]):                -int(0.01*mixframe.shape[0])] = opframe
         
-        cv2.imshow('mixframe',mixframe)
-        cv2.resizeWindow('mixframe',scrw,scrh)
+        mixframe = cv2.rotate(mixframe, cv2.ROTATE_90_COUNTERCLOCKWISE)
+        mixframe = cv2.flip(mixframe,0)
+        mixframe = cv2.cvtColor(mixframe, cv2.COLOR_BGR2RGB)
+        mixframe = pygame.surfarray.make_surface(mixframe)
+        
+        screen.blit(mixframe,(0,0))
+        pygame.display.update()
       else:
         if not imgonly:
           opframe = cv2.resize(opframe,None,fx=immusic.shape[0]/opframe.shape[0],fy=immusic.shape[0]/opframe.shape[0])